Paper Content and Size Adjusted
Attributs of the paper are now controlled and adjusted as the elements are displayed on the paper. If the paper overflows it changes its size accordingly. The positions of the Method, Roles, Resources, Participants, and the Outputs is now set based on the first source element that is visible on the paper.